--- eslint.config.mjs
@@ -1,23 +1,0 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import js from "@eslint/js";
-import { FlatCompat } from "@eslint/eslintrc";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const compat = new FlatCompat({
-  baseDirectory: __dirname,
-  recommendedConfig: js.configs.recommended,
-  allConfig: js.configs.all,
-});
-
-export default [
-  ...compat.extends("next/core-web-vitals"),
-  {
-    rules: {
-      "@next/next/no-html-link-for-pages": "off",
-      "@next/next/no-img-element": "off",
-      "jsx-a11y/alt-text": "off",
-    },
-  },
-];
