+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "type": "module",
   "scripts": {
+    "lint": "eslint src/ app/ chartFormat/",
     "ndev": "next dev -H 0.0.0.0",
     "nbuild": "next build",
     "vdev": "vercel dev --listen 0.0.0.0:8787",
@@ -29,17 +30,20 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.2.0",
-    "@eslint/js": "^9.17.0",
+    "@eslint/js": "^9.19.0",
     "@types/file-saver": "^2.0.7",
     "@types/node": "^22",
     "@types/react": "^19.0.2",
     "@types/react-dom": "^19.0.2",
-    "eslint": "^9.17.0",
+    "eslint": "^9.19.0",
     "eslint-config-next": "^15.1.2",
+    "eslint-plugin-react": "^7.37.4",
+    "globals": "^15.14.0",
     "postcss": "^8",
     "tailwindcss": "^3.4.16",
     "tsx": "^4.19.2",
     "typescript": "^5",
+    "typescript-eslint": "^8.22.0",
     "vercel": "^40.1.0"
   },
   "overrides": {
